Move core functionality from SpecCluster to Cluster
This moves standard functionality from SpecClsuter to the Cluster
superclass.  It also removes the assumption that the Scheduler will be
local to the Cluster class.
